+++ packages/components-react/src/pages/index.tsx
@@ -71,6 +71,8 @@ import { FlagPage } from './generated/Flag';
 import { FlexPage } from './generated/Flex';
 import { Flyout_1Page } from './generated/Flyout_1';
 import { Flyout_2Page } from './generated/Flyout_2';
+import { FlyoutBackdropBlurPage } from './generated/FlyoutBackdropBlur';
+import { FlyoutBackdropShadingPage } from './generated/FlyoutBackdropShading';
 import { FlyoutBasicPage } from './generated/FlyoutBasic';
 import { FlyoutCssVariables_1Page } from './generated/FlyoutCssVariables_1';
 import { FlyoutCssVariables_2Page } from './generated/FlyoutCssVariables_2';
@@ -526,6 +528,16 @@ export const generatedRoutes: RouteType[] = [
     name: 'Flyout 2',
     path: '/flyout-2',
     element: <Flyout_2Page />,
+  },
+  {
+    name: 'Flyout Backdrop Blur',
+    path: '/flyout-backdrop-blur',
+    element: <FlyoutBackdropBlurPage />,
+  },
+  {
+    name: 'Flyout Backdrop Shading',
+    path: '/flyout-backdrop-shading',
+    element: <FlyoutBackdropShadingPage />,
   },
   {
     name: 'Flyout Basic',